+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+### Added
+- Add helm-repo-config to k3d-cluster #109.
+
 ## [1.65.1](https://github.com/cloudogu/ces-build-lib/releases/tag/1.65.1) - 2023-08-17
 
 - `findVulnerabilitiesWithTrivy` schema bug fix